fix bugs

+++ CHelper-Cmd/src/cmd/CHelperCmd.h
@@ -36,7 +36,7 @@ namespace CHelper::Test {
 
     [[maybe_unused]] void test(CHelperCore *core, const std::vector<std::u16string> &commands, bool isTestTime);
 
-    [[maybe_unused]] void test2(const std::filesystem::path &cpackPath, const std::vector<std::u16string> &commands, int times);
+    [[maybe_unused]] void test2(const std::filesystem::path &cpackPath, const std::vector<std::u16string> &commands, size_t times);
 
     [[maybe_unused]] bool writeDirectory(const std::filesystem::path &input, const std::filesystem::path &output, const std::string &fileType);
